+++ config/settings.py
@@ -1,13 +1,19 @@
 """ Flask and Flask extensions configuration file. """
 
-# Timedelta import:
+# Built-in imports:
 from datetime import timedelta
+from pathlib import Path
+
+# CacheLib import:
+from cachelib.file import FileSystemCache
 
 # Minify settings:
 bypass = ["robots_txt", "humans_txt", "sitemap_xml"]
 
 # Session settings:
-SESSION_TYPE = "filesystem"
+sessions_path = Path(__file__).parents[1].resolve() / "flask_session"
+SESSION_TYPE = "cachelib"
+SESSION_CACHELIB = FileSystemCache(threshold=500, cache_dir=sessions_path)
 SESSION_PERMANENT = True
 PERMANENT_SESSION_LIFETIME = timedelta(hours=24)
 SESSION_COOKIE_SECURE = True