+++ docs/CHANGELOG.md
@@ -5,13 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.0] - 2024.03.09
+## [2.0.0] - 2024.03.27
 
 ## Added
 
 - Implemented minification with Flask-Minify.
 - Implemented HTTP security headers with Flask-Talisman.
 - New sources: Pexels, Pixabay, Unsplash.
+- Pylint workflow.
 - SEO files (robots.txt, sitemap.xml).
 - humans.txt.
 - Added more comments.
@@ -22,7 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Eliminated inline JavaScript.
 - HTML restructured for accessibility and SEO.
 - Slow loading "current year" JS swapped out for Jinja/Python variable.
+- ChaceLib settings for new Flask-Session version 0.8.0.
 - Updated docs.
+- Updated requirements.txt.
 
 ### Fixed